Allow verity commands on files opened with O_PATH
Verity fs opens symlinks with O_PATH | O_NOFOLLOW to enable/measure
verity on symlinks.

Currently vfs ioctl does not allow input fd opened with O_PATH. Added
conditions to allow verity commands to bypass the check.

Also The existing check in vfs OpenAt() to bypass special case for
verity does work since it's checking the mount of the root dentry.
Changed it to check for the dentry that repesents the actual path.

PiperOrigin-RevId: 381977463